fix(audit): corrections critiques/majeures de l'audit complet
- Auth: must_change_password applique cote serveur, JWT enrichi en permissions/categorie a jour a chaque requete
- Routes /utilisateurs et /admin/reset reservees aux Admin (adminOnly)
- Validation prix/quantite/montant_paye dans ventes et achats
- prix_achat non tronque (parseFloat au lieu de parseInt)
- Numerotation de facture atomique (advisory lock)
- Verification de stock atomique a la vente (FOR UPDATE, anti-survente)
- Suppression des fuites de messages d'erreur SQL bruts
- decodeURIComponent manquant dans generateRecu
- Securite HTTP (headers) + rate limiting sur /auth/login
- Index DB manquants (entreprise_id, dates, FKs)
- Frontend: cleanup useEffect Articles.jsx, centralisation logique facture reglee/marge, correction arrondi repartition paiement Achats.jsx

+++ backend/src/controllers/articlesController.js
@@ -79,7 +79,9 @@ async function create(req, res) {
       const date  = new Date().toISOString().split("T")[0];
       const mois  = MOIS[new Date(date).getMonth()];
       const annee = new Date(date).getFullYear();
-      const prixUnitaire = parseInt(prix_achat) || 0;
+      // prix_achat est un NUMERIC en base (peut comporter des décimales) :
+      // utiliser parseFloat, pas parseInt, sous peine de tronquer (ex: 1499.99 → 1499).
+      const prixUnitaire = parseFloat(prix_achat) || 0;
       const montantTotal = prixUnitaire * qteInitiale;
 
       await client.query(
